refactor: Controller로 요청 url 분기 처리
