+++ src/main/java/controller/Controller.java
@@ -1,0 +1,8 @@
+package controller;
+
+import util.HttpRequest;
+import util.HttpResponse;
+
+public interface Controller {
+    void service(HttpRequest request, HttpResponse response);
+}
